requirements_test.txt
autopep8==1.0.3
coverage==3.6

django-jenkins==0.14.1
django-nose==1.2
django-pytest==0.2.0
docutils==0.12
factory-boy==2.2.1
flake8==3.2.1

flower
HTTPretty==0.8.3
Jinja2==2.7.3
mock==1.0.1
nose==1.3.0
pep8==1.5.7

pyflakes==1.2.3
pycodestyle==2.0.0
clonedigger==1.1.0
shortuuid==0.5.0
Pygments==1.6
PyHamcrest==1.9.0
pytest==2.2.4
pytest-cov==1.6
PyYAML==3.10
Sphinx==1.2.2
sphinx-rtd-theme==0.1.6
sure==1.2.7
